2 tests still not working
