+++ src/components/ProgressBar/index.spec.js
@@ -1,8 +1,14 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
+import configureStore from 'redux-mock-store'
+
 import ProgressBar from '.'
-import StyledAnsweredQuestionNumber from '.'
+import { getQuestionNumber } from './QuestionNumber'
+
+import { initialState } from '../../state'
+
+const mockStore = configureStore()
 
 describe('components:ProgressBar', function () {
   it('renders the ProgressBar and CSS properties properly', () =>
@@ -10,17 +16,19 @@ describe('components:ProgressBar', function () {
       toJson(shallow(<ProgressBar>ProgressBar</ProgressBar>))
     ).toMatchSnapshot())
 
-  it('renders the questionNumber CSS properties properly', () => {
+  it(`renders the questionNumber CSS properties properly`, () => {
     const questionNumber = 1
-    expect(
-      toJson(
-        shallow(
-          <StyledAnsweredQuestionNumber
-            key={questionNumber}
-            onClick={clickHandler}
-          />
-        ).dive()
-      )
-    ).toMatchSnapshot()
+    const activeQuestion = 3
+    const store = mockStore(initialState)
+
+    store.dispatch = jest.fn()
+
+    const wrapper = shallow(
+      <ProgressBar store={store} activeQuestion={getQuestionNumber} />
+    )
+
+    wrapper.props().getQuestionNumber()
+
+    expect(store.dispatch).toHaveBeenCalledWith({})
   })
 })
