+++ src/components/styled/PreviousQuestions/index.spec.js
@@ -5,10 +5,12 @@ import PreviousQuestions from '.'
 
 describe('components:styled:PreviousQuestions', () => {
   it('renders the PreviousQuestions CSS properties properly', () => {
-    expect(toJson(shallow(<PreviousQuestions />))).toMatchSnapshot()
+    expect(
+      toJson(shallow(<PreviousQuestions>PreviousQuestions</PreviousQuestions>))
+    ).toMatchSnapshot()
   })
 
-  it('renders the PreviousQuestions properly with enalbed prop', () => {
+  it('renders the PreviousQuestions properly with enabled prop', () => {
     expect(toJson(shallow(<PreviousQuestions enabled />))).toMatchSnapshot()
   })
 })
